+++ pkg/app/services/publishing/removeVersion/main.go
@@ -3,11 +3,14 @@ package removeVersion
 import (
 	"creatif/pkg/app/auth"
 	"creatif/pkg/app/domain/published"
+	"creatif/pkg/app/services/events"
 	pkg "creatif/pkg/lib"
 	"creatif/pkg/lib/appErrors"
+	"creatif/pkg/lib/constants"
 	"creatif/pkg/lib/logger"
 	"creatif/pkg/lib/storage"
 	"fmt"
+	"os"
 )
 
 type Results struct {
@@ -42,8 +45,18 @@ func (c Main) Authorize() error {
 }
 
 func (c Main) Logic() (*struct{}, error) {
+	var version published.Version
+	if res := storage.Gorm().Raw(fmt.Sprintf("SELECT id, name FROM %s WHERE project_id = ? AND id = ?", (published.Version{}).TableName()), c.model.ProjectID, c.model.ID).Scan(&version); res.Error != nil {
+		return nil, appErrors.NewApplicationError(res.Error)
+	}
+
 	if res := storage.Gorm().Exec(fmt.Sprintf("DELETE FROM %s WHERE project_id = ? AND id = ?", (published.Version{}).TableName()), c.model.ProjectID, c.model.ID); res.Error != nil {
 		return nil, appErrors.NewApplicationError(res.Error)
+	}
+
+	path := fmt.Sprintf("%s/%s/%s", constants.PublicDirectory, c.model.ProjectID, version.Name)
+	if err := os.RemoveAll(path); err != nil {
+		events.DispatchEvent(events.NewPublicDirectoryNotRemoved(path, "", c.model.ProjectID))
 	}
 
 	return nil, nil
